Bill call usage by seconds and sync expired subscription status.
Use fractional minutes for fair deductions, auto-mark balances expired when admin changes the end date, and add decimal migrations for balance fields.

+++ apps/calls/models.py
@@ -64,7 +64,9 @@ class CallSession(models.Model):
         default=False,
         verbose_name="مقابلة إدارة (مكالمة معلم جديد)",
     )
-    minutes_charged = models.PositiveIntegerField(
+    minutes_charged = models.DecimalField(
+        max_digits=12,
+        decimal_places=4,
         null=True,
         blank=True,
         verbose_name="دقائق مخصومة من الرصيد",
